+++ src/main/java/com/itechart/stockOnline/dao/RoomDao.java
@@ -12,16 +12,16 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
-public interface RoomDao extends JpaRepository<Room, Long>, JpaSpecificationExecutor{
-    Optional<Room> findById(Long id);
+public interface RoomDao extends JpaRepository<Room, Integer>, JpaSpecificationExecutor{
+    Optional<Room> findById(Integer id);
     Page<Room> findAll(Pageable pageable);
 
     @Modifying
     @Query("delete from Room s where s.id in ?1")
-    int deleteByIdIn(Collection<Long> ids);
+    int deleteByIdIn(Collection<Integer> ids);
 
     @Query("select s from Room s where s.id in ?1")
-    Stream<Room> findAllByIdIn(Collection<Long> ids);
+    Stream<Room> findAllByIdIn(Collection<Integer> ids);
 
     Set<Room> findAllByStockId(Long stockId);
